Integrated VICE's Z80 emu module into AppleWin (Still have Z80Em module too)

+++ AppleWin/source/Z80Em/z80io.h
@@ -23,7 +23,7 @@ void Z80_Out (BYTE Port,BYTE Value);
 /****************************************************************************/
 /* Read a byte from given memory location                                   */
 /****************************************************************************/
-unsigned Z80_RDMEM(DWORD A);
+BYTE Z80_RDMEM(DWORD A);
 
 /****************************************************************************/
 /* Write a byte to given memory location                                    */
